Annotate some `toArray(T[])` methods.
There's no perfect way to annotate these. The most technically correct
way would be to require an array of a `@Nullable` type. That
accommodates both collections that contain nulls and collections that
are smaller than the input array (which then write a null terminator
into the array). But that signature would be kind of a pain for normal
usage, so let's try this.