Annotate AtomicReference{,Array} for nullness.
There is some question about what the best way to annotate these classes
is. The annotations in this commit produce an unsound API, since the
no-arg `new AtomicReference<>()` constructor can produce a reference
with a null value even for a type that shouldn't permit nulls. Still,
this seems like the least bad option.