Require Class to be instantiated with a non-nullable type.
This is similar to the previous commit. The stakes are perhaps a little
higher in this case:

Without this commit, it is easy to imagine that users will end up with a
`Class<T>` and a `Class<@Nullable T>`, which would be inconvertible
types under JSpecify. _Plus_, it would be a shame for `newInstance()` to
return a nullable value (which is would if we permitted nullable
instantiations because we don't have a `@NonNull` annotation on its
return type).

On the other hand, with this commit, it's also easy to imagine that the
inability to write `Class<T>` when `<T extends @Nullable Object>` will
be a pain. (But users can at least suppress and get the right behavior
in that case.)