Experiment with a different signature for `Map.compute*`.
Although those methods' functions are _allowed_ to return `null`, I'm
wondering if they would be more practically useful without that ability.
Without it, we can restrict the return type to `V`. Given that the
`compute*` methods are sometimes used specifically to ensure that the
returned value is not null, this may be a pragmatic compromise.

Then again, we can imagine tools that would outright _forbid_ using
`null` in that case (unless the caller casts the whole `Map` to have a
nullable value type...). That would be unfortunate, and maybe I'm
overstepping. But, to find out, let's try the less flexible signature.

(This is of course all an argument for `@PolyNull`. Or maybe this is one
of those cases in which we'd really want a _variant_ of `@PolyNull`? I
forget.)

Oh, and some of these signatures were just outright wrong. This commit
fixes those. So if we someday revert this commit, we'll want to restore
those changes.

TODO(cpovirk): Look at the upstream signatures to see if they're
consistent and correct.