Experiment with a narrower signature for _some_ `Thread` constructors.
In all cases, it's possible to omit the `Runnable` parameter instead of
passing `null`:

- `new Thread(null)` => `new Thread()`
- `new Thread(null, name)` => `new Thread(name)`
- `new Thread(group, null, name)` => `new Thread(group, name)`

Generally speaking, users of `Thread` should know whether they are
creating a `Thread` that overrides `run()` (in which case they'll want
to omit the `Runnable` parameter) or one that does not (in which case
it's important that they pass a non-null `Runnable` so that their thread
actually does something).

`Thread` has some other constructors that accept a `Runnable`. However,
those accept a `stackSize` (and sometimes `inheritThreadLocals`), and
the class provides no way for callers to provide those values without
also passing a `Runnable`. Thus, we keep those constructors' `Runnable`
parameters as `@Nullable` to accommodate `Thread` subclasses that
override `run()` and call those constructors.

I can imagine that the narrower signatures will cause problems for
general-purpose subclasses of `Thread`. And I can also imagine that they
won't catch meaningful problems in practice. But let's give it a try....