Method.invoke: Permit a null array of args.
My hope had been to get away without this: Callers can always pass an
*empty* array, and in fact they can often just pass no args, since the
method changed to a varargs method way back when. Given that, forbidding
null seemed like a cheap way to catch anyone who accidentally passed an
uninitialized field or `(Object[]) null` instead of `(Object) null`.

Granted, I was a little worried that there would be too much existing
code that we'd require needless updates to. But I wanted to at least
find out. And I hoped that we could reduce the busy work by updating
callers automatically.

In the end, though, I'm seeing another significant drawback: One common
reason to call Method.invoke is inside an InvocationHandler. And
InvocationHandler accepts an array of args -- which is *always* null in
the case of a no-arg method. So, if we forbid a null array of args in
Method.invoke, we prevent callers from being able to pass through the
InvocationHandler args directly.

So I'm giving in.

(I still haven't given in for Constructor, since InvocationHandler isn't
as likely to be used with that (though it could be). But it seems likely
that I'll reverse course because it will be harder to justify the
automation to update callers when we'd be touching only callers of
Constructor and not callers of Method.)