Simplify signature of `Arrays.copyOf`/`copyOfRange`.
We already say that the output array can contain nulls, so there's no
difference between the result type you get from
`Arrays.<Foo>copyOf(...)` and the result type you get from
`Arrays.<@Nullable Foo>copyOf(...)`.

(Checkers will ideally provide a narrower return type when possible. But
they can do that with either signature.)

Plus, this simplification eliminates the suggestion that it's possible
to have a `Class<@Nullable Foo[]>`. For all I know, that might be
possible. (Note that that's different from the definitely
should-be-impossible `Class<Foo @Nullable []>`.) But it's something that
I wouldn't want to encourage unless we had a good reason.

Or, alternative viewpoint: Maybe _every_ `Class` for an array type
_should_ be a `Class<@Nullable Foo[]>`. That way, `clazz.cast(...)`
can't convert a `@Nullable Foo[]` to a `Foo[]`, as it would if we let
users have a `Class<Foo[]>`. Still, if we want to do that, then the
right way is going to be to make this method accept a `Class<? extends
@Nullable T[]>` so as to _force_ that type. And then there would be the
question of whether we can force `Class` objects for array types to take
that form in general, such as from `Foo[].class`....