Annotate much of `java.time` for nullness.
Adapted from upstream PR 72.

I decided not to add the `@Nullable` annotations to the various
`isSupported` methods (and also `Year.isValidMonthDay`), since it's not
clear that anyone would pass `null` intentionally. However, we should
revisit this the moment that we learn of anyone who is trying to do so.

---

Also, as an unrelated aside: I looked at recent commits to upstream when
I updated our tracking branch to
https://github.com/jspecify/jdk/commit/3d55d8c9f7a723241ce8a1d22ba48f9b8a46bb55.
As best I can tell, we have picked up all upstream changes except for:

- src/java.base/share/classes/java/util/Properties.java, which was
  changed in 20c1642dd0a7b14061f52eeac10d6f89adfd0254. The most relevant
  change was to `merge`, but I'm trying not to get into the various
  functional methods on `Map` and its implementations yet.
- src/java.desktop/share/classes/java/awt/Window.java, which was changed
  in 62877f06ccae92c79fd351a63bf0b546147a0e6f. It doesn't seem worth
  getting into AWT until we have a more automated process for keeping up
  to date with upstream.