Make type arguments to `Class` and `Optional` non-nullable.
We've declared those classes' type parameters as having a bound of
`@NonNull Object`, as discussed in
https://github.com/jspecify/jspecify/issues/78. This PR changes type
arguments inside the JDK to be inside that bound.

eisop and typetools, by contrast, use a bound of `@Nullable Object`,
which they "undo" with `@NonNull T` on methods like `Class.newInstance`.
Thus, they don't need or want a change like this one. (That said, they
don't _universally_ use `@Nullable Object` as the bound in "cases like
this," so it's possible that they'd feel that a particular API, maybe
`ServiceLoader`, would be simpler with a non-nullable bound. Or, if not,
they may wish to use `@NonNull S` at the use sites.)

(The complicated API in this PR is `ServiceLoader`. I do see some docs
that suggest that `ServiceLoader` rejects any attempt to return a `null`
service instance, and that appears to be backed up by the implementation
I traced through to `ProviderImpl.get`.)