Annotate `Map.merge` with `@Nullable` on the method and function return type.
We had been trying to get by without, requiring users to suppress where
needed. But our philosophy has shifted more toward "If you want to be
able to put null in or get null out, then it needs to be `@Nullable`."

This change brings us closer to upstream, which uses `@PolyNull`.