Allow `Node.setNodeValue` to be called with a null argument.
The specification is a bit unclear, but it  seems to imply that null is OK. The
implementation also behaves that way. Also, having `getNodeValue` return
`@Nullable String` but `setNodeValue` take just `String` means that Kotlin
doesn't recognize them as a get/set pair so you can't write `node.nodeValue =
whatever`.