Make `enumerate` accept `@Nullable Thread[]` (ditto `ThreadGroup`).
Since the method only writes to the array, it is fine with either a
nullable element type or a non-nullable element type. (I'm operating
under the usual assumption that tools treat arrays as covariant. For
purposes of Checker Framework users who set `-AinvariantArrays`, we
could consider using `@PolyNull` instead in eisop.)

This change is prompted by a Kotlin caller that passes an array of
`Thread?` because Kotlin won't allow an uninitialized array of `Thread`.
In Java, I would expect most users to allocate an uninitialized array of
`[@NonNull] Thread` and pass that.

(This is a pretty weird API, though, since it requires you to pass in an
array that you pre-size for the number of threads you expect and so you
might overshoot and end up (I assume; the docs are not clear) with nulls
in the result....)

(#28)