Annotate `computeIfPresent` and `compute` more like the Checker Framework and Kotlin do.
In contrast, for _`computeIfAbsent`_, I've kept the signature _without_
`@Nullable` on the function result type and method return type. This is
far more convenient for most callers, and it just might not cause
problems for anyone: Callers of `computeIfAbsent` might _always_ provide
a function that elects to insert a value (as expressed by returning a
non-null value).

I think this gives us the signatures that we'll want for the long term
for `Map`, though of course we'd revisit if we chose to include
`@PolyNull`.

Previously:

- https://github.com/jspecify/jdk/pull/14#issuecomment-1642254198
- https://github.com/jspecify/jdk/pull/19#issuecomment-1622173811

Note that this PR excludes `java.util.Properties` because there are a
number of other changes that I'm going to make to it (and then
upstream).