Annotate the new `repeat` API on `StringBuilder` and friends. (#84)
I'm not sure that annotation in `AbstractStringBuilder` matter much, but
I'm not ruling it out that an expression could have that implicit type
if you wrote something like `b ? new StringBuilder() : new
StringBuffer()`. And anyway, it was nice to keep `AbstractStringBuilder`
up to date while having reference to its implementation so that I could
then easily cross-reference between it and the subclasses.